refactor: extract command processing workflow
